# docs/swagger.yaml
basePath: /
definitions:
  entity.AllSetting:
    properties:
      datepicker:
        description: Date picker format
        type: string
      expireDiff:
        description: Expiration warning threshold in days
        type: integer
      externalTrafficInformEnable:
        description: Enable external traffic reporting
        type: boolean
      externalTrafficInformURI:
        description: URI for external traffic reporting
        type: string
      ldapAutoCreate:
        type: boolean
      ldapAutoDelete:
        type: boolean
      ldapBaseDN:
        type: string
      ldapBindDN:
        type: string
      ldapDefaultExpiryDays:
        type: integer
      ldapDefaultLimitIP:
        type: integer
      ldapDefaultTotalGB:
        type: integer
      ldapEnable:
        description: LDAP settings
        type: boolean
      ldapFlagField:
        description: Generic flag configuration
        type: string
      ldapHost:
        type: string
      ldapInboundTags:
        type: string
      ldapInvertFlag:
        type: boolean
      ldapPassword:
        type: string
      ldapPort:
        type: integer
      ldapSyncCron:
        type: string
      ldapTruthyValues:
        type: string
      ldapUseTLS:
        type: boolean
      ldapUserAttr:
        description: e.g., mail or uid
        type: string
      ldapUserFilter:
        type: string
      ldapVlessField:
        type: string
      pageSize:
        description: UI settings
        type: integer
      remarkModel:
        description: Remark model pattern for inbounds
        type: string
      sessionMaxAge:
        description: Session maximum age in minutes
        type: integer
      subAnnounce:
        description: Subscription announce
        type: string
      subCertFile:
        description: SSL certificate file for subscription server
        type: string
      subDomain:
        description: Domain for subscription server validation
        type: string
      subEnable:
        description: Subscription server settings
        type: boolean
      subEnableRouting:
        description: Enable routing for subscription
        type: boolean
      subEncrypt:
        description: Encrypt subscription responses
        type: boolean
      subJsonEnable:
        description: Enable JSON subscription endpoint
        type: boolean
      subJsonFragment:
        description: JSON subscription fragment configuration
        type: string
      subJsonMux:
        description: JSON subscription mux configuration
        type: string
      subJsonNoises:
        description: JSON subscription noise configuration
        type: string
      subJsonPath:
        description: Path for JSON subscription endpoint
        type: string
      subJsonRules:
        type: string
      subJsonURI:
        description: JSON subscription server URI
        type: string
      subKeyFile:
        description: SSL private key file for subscription server
        type: string
      subListen:
        description: Subscription server listen IP
        type: string
      subPath:
        description: Base path for subscription URLs
        type: string
      subPort:
        description: Subscription server port
        type: integer
      subProfileUrl:
        description: Subscription profile URL
        type: string
      subRoutingRules:
        description: Subscription global routing rules (Only for Happ)
        type: string
      subShowInfo:
        description: Show client information in subscriptions
        type: boolean
      subSupportUrl:
        description: Subscription support URL
        type: string
      subTitle:
        description: Subscription title
        type: string
      subURI:
        description: Subscription server URI
        type: string
      subUpdates:
        description: Subscription update interval in minutes
        type: integer
      tgBotAPIServer:
        description: Custom API server for Telegram bot
        type: string
      tgBotBackup:
        description: Enable database backup via Telegram
        type: boolean
      tgBotChatId:
        description: Telegram chat ID for notifications
        type: string
      tgBotEnable:
        description: Telegram bot settings
        type: boolean
      tgBotLoginNotify:
        description: Send login notifications
        type: boolean
      tgBotProxy:
        description: Proxy URL for Telegram bot
        type: string
      tgBotToken:
        description: Telegram bot token
        type: string
      tgCpu:
        description: CPU usage threshold for alerts
        type: integer
      tgLang:
        description: Telegram bot language
        type: string
      tgRunTime:
        description: Cron schedule for Telegram notifications
        type: string
      timeLocation:
        description: Security settings
        type: string
      trafficDiff:
        description: Traffic warning threshold percentage
        type: integer
      twoFactorEnable:
        description: Enable two-factor authentication
        type: boolean
      twoFactorToken:
        description: Two-factor authentication token
        type: string
      webBasePath:
        description: Base path for web panel URLs
        type: string
      webCertFile:
        description: Path to SSL certificate file for web server
        type: string
      webDomain:
        description: Web server domain for domain validation
        type: string
      webKeyFile:
        description: Path to SSL private key file for web server
        type: string
      webListen:
        description: Web server settings
        type: string
      webPort:
        description: Web server port number
        type: integer
    type: object
  entity.Msg:
    properties:
      msg:
        description: Response message text
        type: string
      obj:
        description: Optional data object
      success:
        description: Indicates if the operation was successful
        type: boolean
    type: object
...
host: localhost:2053
info:
  contact: {}
  description: API documentation for 3X-UI Panel (Master-Slave Cluster)
  title: 3X-UI Cluster API
  version: "1.0"
paths:
...
  /panel/api/outbounds/del/{id}:
    post:
      description: Deletes an outbound configuration from a slave
      parameters:
      - description: Outbound index
        in: path
        name: id
        required: true
        type: integer
      - description: Slave ID
        in: query
        name: slaveId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Delete outbound
      tags:
      - Outbounds
  /panel/api/outbounds/list:
    get:
      description: Returns outbound configurations for a specific slave
      parameters:
      - description: Slave ID
        in: query
        name: slaveId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: List outbounds
      tags:
      - Outbounds
...
  /panel/api/routing/del/{id}:
    post:
      description: Deletes a routing rule from a slave
      parameters:
      - description: Rule index
        in: path
        name: id
        required: true
        type: integer
      - description: Slave ID
        in: query
        name: slaveId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Delete routing rule
      tags:
      - Routing
  /panel/api/routing/list:
    get:
      description: Returns routing rules for a specific slave
      parameters:
      - description: Slave ID
        in: query
        name: slaveId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: List routing rules
      tags:
      - Routing
...
  /panel/api/setting/update:
    post:
      consumes:
      - application/json
      description: Updates all panel settings
      parameters:
      - description: Settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/entity.AllSetting'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Update settings
      tags:
      - Settings
...
  /panel/api/xray/getOutboundsTraffic:
    get:
      description: Returns traffic statistics for outbound connections
      parameters:
      - description: Slave ID (-1 for all)
        in: query
        name: slaveId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Get outbound traffic
      tags:
      - XraySettings
...
  /panel/api/xray/warp/{action}:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Handles Warp registration, config, data, and license operations
      parameters:
      - description: Action (data/del/config/reg/license)
        in: path
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Warp operations
      tags:
      - XraySettings
securityDefinitions:
  SessionAuth:
    in: cookie
    name: 3x-ui
    type: apiKey
swagger: "2.0"