# web/controller/xray_setting.go
package controller

import (
	"encoding/json"
	"fmt"
	"strconv"

	"github.com/mhsanaei/3x-ui/v2/database/model"
	"github.com/mhsanaei/3x-ui/v2/logger"

	"github.com/mhsanaei/3x-ui/v2/web/service"

	"github.com/gin-gonic/gin"
)

// XraySettingController handles Xray configuration and settings operations.
type XraySettingController struct {
	XraySettingService  service.XraySettingService
	SlaveSettingService service.SlaveSettingService
	SettingService      service.SettingService
	InboundService      service.InboundService
	OutboundService     service.OutboundService
	XrayService         service.XrayService
	WarpService         service.WarpService
}

// NewXraySettingController creates a new XraySettingController and initializes its routes.
func NewXraySettingController(g *gin.RouterGroup) *XraySettingController {
	a := &XraySettingController{}
	a.initRouter(g)
	return a
}

// initRouter sets up the routes for Xray settings management.
func (a *XraySettingController) initRouter(g *gin.RouterGroup) {
	g.GET("/getDefaultJsonConfig", a.getDefaultXrayConfig)
	g.GET("/getOutboundsTraffic", a.getOutboundsTraffic)
	g.GET("/getXrayResult", a.getXrayResult)

	g.POST("/", a.getXraySetting)
	g.POST("/warp/:action", a.warp)
	g.POST("/update", a.updateSetting)
	g.POST("/resetOutboundsTraffic", a.resetOutboundsTraffic)

}

// getXraySetting retrieves the Xray configuration template, inbound tags, and outbound test URL.
// @Summary Get Xray settings
// @Description Returns Xray configuration for a specific slave
// @Tags XraySettings
// @Accept json
// @Produce json
// @Success 200 {object} entity.Msg
// @Router /panel/api/xray/ [post]
func (a *XraySettingController) getXraySetting(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		jsonMsg(c, I18nWeb(c, "error"), err)
		return
	}

	// Extract slaveId from request body
	slaveIdFloat, ok := req["slaveId"].(float64)
	if !ok || int(slaveIdFloat) <= 0 {
		jsonMsg(c, "请选择一个Slave节点", fmt.Errorf("slaveId is required"))
		return
	}
	slaveId := int(slaveIdFloat)
	
	// Use SlaveSettingService to get per-slave configuration
	xraySetting, err := a.SlaveSettingService.GetXrayConfigForSlave(slaveId)
	if err != nil {
		jsonMsg(c, I18nWeb(c, "pages.settings.toasts.getSettings"), err)
		return
	}
	inboundTags, err := a.InboundService.GetInboundTags()
	if err != nil {
		jsonMsg(c, I18nWeb(c, "pages.settings.toasts.getSettings"), err)
		return
	}
	xrayResponse := map[string]interface{}{
		"xraySetting":     json.RawMessage(xraySetting),
		"inboundTags":     json.RawMessage(inboundTags),
	}
	result, err := json.Marshal(xrayResponse)
	if err != nil {
		jsonMsg(c, I18nWeb(c, "pages.settings.toasts.getSettings"), err)
		return
	}
	jsonObj(c, string(result), nil)
}

// updateSetting updates the Xray configuration settings.
// @Summary Update Xray settings
// @Description Updates the Xray configuration for a specific slave
// @Tags XraySettings
// @Accept json
// @Produce json
// @Success 200 {object} entity.Msg
// @Router /panel/api/xray/update [post]
func (a *XraySettingController) updateSetting(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		jsonMsg(c, I18nWeb(c, "error"), err)
		return
	}

	// Extract slaveId from request body
	slaveIdFloat, ok := req["slaveId"].(float64)
	if !ok || int(slaveIdFloat) <= 0 {
		jsonMsg(c, "请选择一个Slave节点", fmt.Errorf("slaveId is required"))
		return
	}
	slaveId := int(slaveIdFloat)
	
	// Use SlaveSettingService to save per-slave configuration
	xraySetting, ok := req["xraySetting"].(string)
	if !ok {
		jsonMsg(c, I18nWeb(c, "error"), fmt.Errorf("xraySetting is required"))
		return
	}
	
	// Validate config first
	if err := a.XraySettingService.CheckXrayConfig(xraySetting); err != nil {
		jsonMsg(c, I18nWeb(c, "pages.settings.toasts.modifySettings"), err)
		return
	}
	
	err := a.SlaveSettingService.SaveXrayConfigForSlave(slaveId, xraySetting)
	if err == nil {
		go func() {
			slaveService := service.SlaveService{}
			if err := slaveService.PushConfig(slaveId); err != nil {
				logger.Warningf("XraySettingController: failed to push config to slave %d: %v", slaveId, err)
			}
		}()
	}
	jsonMsg(c, I18nWeb(c, "pages.settings.toasts.modifySettings"), err)
}

// getDefaultXrayConfig retrieves the default Xray configuration.
// @Summary Get default Xray config
// @Description Returns the default Xray configuration template
// @Tags XraySettings
// @Produce json
// @Success 200 {object} entity.Msg
// @Router /panel/api/xray/getDefaultJsonConfig [get]
func (a *XraySettingController) getDefaultXrayConfig(c *gin.Context) {
	defaultJsonConfig, err := a.SettingService.GetDefaultXrayConfig()
	if err != nil {
		jsonMsg(c, I18nWeb(c, "pages.settings.toasts.getSettings"), err)
		return
	}
	jsonObj(c, defaultJsonConfig, nil)
}

// getXrayResult retrieves the current Xray service result.
// @Summary Get Xray result
// @Description Returns the current Xray service running result
// @Tags XraySettings
// @Produce json
// @Success 200 {object} entity.Msg
// @Router /panel/api/xray/getXrayResult [get]
func (a *XraySettingController) getXrayResult(c *gin.Context) {
	jsonObj(c, a.XrayService.GetXrayResult(), nil)
}

// warp handles Warp-related operations based on the action parameter.
// @Summary Warp operations
// @Description Handles Warp registration, config, data, and license operations
// @Tags XraySettings
// @Accept x-www-form-urlencoded
// @Produce json
// @Param action path string true "Action (data/del/config/reg/license)"
// @Success 200 {object} entity.Msg
// @Router /panel/api/xray/warp/{action} [post]
func (a *XraySettingController) warp(c *gin.Context) {
	action := c.Param("action")
	var resp string
	var err error
	switch action {
	case "data":
		resp, err = a.WarpService.GetWarpData()
	case "del":
		err = a.WarpService.DelWarpData()
	case "config":
		resp, err = a.WarpService.GetWarpConfig()
	case "reg":
		skey := c.PostForm("privateKey")
		pkey := c.PostForm("publicKey")
		resp, err = a.WarpService.RegWarp(skey, pkey)
	case "license":
		license := c.PostForm("license")
		resp, err = a.WarpService.SetWarpLicense(license)
	}

	jsonObj(c, resp, err)
}

// getOutboundsTraffic retrieves the traffic statistics for outbounds.
// @Summary Get outbound traffic
// @Description Returns traffic statistics for outbound connections
// @Tags XraySettings
// @Produce json
// @Param slaveId query int false "Slave ID (-1 for all)"
// @Success 200 {object} entity.Msg
// @Router /panel/api/xray/getOutboundsTraffic [get]
func (a *XraySettingController) getOutboundsTraffic(c *gin.Context) {
	slaveIdStr := c.DefaultQuery("slaveId", "-1")
	slaveId, _ := strconv.Atoi(slaveIdStr)
	
	var outboundsTraffic []*model.OutboundTraffics
	var err error
	
	if slaveId == -1 {
		outboundsTraffic, err = a.OutboundService.GetOutboundsTraffic()
	} else {
		outboundsTraffic, err = a.OutboundService.GetOutboundsTrafficForSlave(slaveId)
	}
	
	if err != nil {
		jsonMsg(c, I18nWeb(c, "pages.settings.toasts.getOutboundTrafficError"), err)
		return
	}
	jsonObj(c, outboundsTraffic, nil)
}

// resetOutboundsTraffic resets the traffic statistics for the specified outbound tag.
// @Summary Reset outbound traffic
// @Description Resets traffic counters for a specific outbound on a slave
// @Tags XraySettings
// @Accept json
// @Produce json
// @Success 200 {object} entity.Msg
// @Router /panel/api/xray/resetOutboundsTraffic [post]
func (a *XraySettingController) resetOutboundsTraffic(c *gin.Context) {
	var req map[string]interface{}
	if err := c.ShouldBindJSON(&req); err != nil {
		jsonMsg(c, I18nWeb(c, "error"), err)
		return
	}

	// Extract slaveId from request body
	slaveIdFloat, ok := req["slaveId"].(float64)
	if !ok || int(slaveIdFloat) <= 0 {
		jsonMsg(c, "请选择一个Slave节点", fmt.Errorf("slaveId is required"))
		return
	}
	slaveId := int(slaveIdFloat)
	
	tag, _ := req["tag"].(string)
	
	err := a.OutboundService.ResetOutboundTrafficForSlave(slaveId, tag)
	if err != nil {
		jsonMsg(c, I18nWeb(c, "pages.settings.toasts.resetOutboundTrafficError"), err)
		return
	}
	jsonObj(c, "", nil)
}